+++ config/responsive-images.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    /**
+     * If the generate images job is being queued, specify the name of the target queue
+     * Falls back to the 'default' queue
+     */
+    'queue' => 'default',
+];